chore: clippy round across workspace + all targets
Sweep clippy warnings the codebase had accumulated since the last
pass. Mostly machine-applicable auto-fixes (`cargo clippy --fix`)
plus a handful of manual ones for the lints clippy can't safely
auto-apply.

Auto-fixed (mechanical):
- needless borrows / reborrows
- needless `as` casts where source and target type matched
- `map_or` → cleaner forms where MSRV permits
- `let-else` → `?` operator in an `Option` builder
- `repeat().take(N)` → `repeat_n(_, N)`
- two `assert!(constant)` → `const { assert!(...) }`
- `>= y + 1` → `> y` in the time-monotonicity test
- `items_after_test_module` reorder in src/window/imp.rs — the
  cairo helpers + the three `impl MeditateWindow` blocks moved
  above the `#[cfg(test)] mod sync_status_tests` so the test
  module is the file's last item.

Manual:
- `meditate-core/src/db/device.rs` Lamport doc-comment: `+ 1.
  Returns` was parsed as a markdown numbered list. Restructured
  the sentence so no `\d+\.` survives in prose.
- `meditate-core/src/sync/webdav.rs` rate-limit doc gained a blank
  line so the trailing prose stops attaching to the bullet list.
- `src/timer/imp.rs` orphan doc-comment (a stub from a moved fn)
  deleted; nearby session-builder doc separator switched to `///`
  so clippy stops flagging the empty-line-after-doc.
- `meditate-core/src/data_io.rs` extracted
  `pub type ImportedRow = (i64, u32, SessionMode, Option<String>,
  usize)` so the import second-pass signature stops triggering
  the very-complex-type lint.
- `src/preferences.rs` keychain-error log: `.inspect_err(|e| log)`
  reverted to a `match` since `inspect_err` is stable since 1.76
  and the crate pins MSRV at 1.75. The clippy auto-fix had
  introduced this; the new `clippy.toml` msrv pin (below)
  prevents the regression from recurring.
- `meditate-core/src/date_math.rs` `i.is_multiple_of(7)` reverted
  to `i % 7 == 0` (the auto-fix used a 1.87 API).
- `meditate-core/src/db/sessions.rs` MIN_DATE saturation test:
  `assert!(n >= 0)` against a `u32` was a tautology (clippy errored
  on it); the surrounding `.unwrap()` already is the "doesn't
  panic" assertion the test intends.

`clippy.toml` added:
- `msrv = "1.75"` so clippy stops proposing fixes that require a
  newer Rust than the project's pinned floor.
- `too-many-arguments-threshold = 12` so the GTK widget-glue
  helpers (`rebuild_chooser_rows`, `present_*_dialog`,
  `wire_boxbreath_phase`, etc.) stop triggering the lint —
  refactoring each into a config struct would shift verbosity
  to the struct definition without making the call sites clearer.

Workspace test count unchanged at 1019 + 62 + 3.

+++ src/stats/imp.rs
@@ -325,13 +325,13 @@ impl StatsView {
         let (title, body) = match &key {
             InsightKey::CurrentStreak { days, is_record, best } => {
                 let body = if *is_record {
-                    ngettext("1 day — new record", "{n} days — new record", *days as u32)
+                    ngettext("1 day — new record", "{n} days — new record", *days )
                         .replace("{n}", &days.to_string())
                 } else if *best > *days {
                     ngettext(
                         "1 day · best was {best}",
                         "{n} days · best was {best}",
-                        *days as u32,
+                        *days ,
                     )
                         .replace("{n}", &days.to_string())
                         .replace("{best}", &best.to_string())
